Assignment-2/Problem_Statement_1/Python/Method_1/pyproject.toml
[project]
name = "method-1"
version = "0.1.0"
description = "Weather App using CLI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.32.4",
    "python-dotenv>=1.0.0",
]
